add payment VNpay 6/5/2025

+++ src/main/java/com/elearning/haui/domain/dto/CourseRepone.java
@@ -15,7 +15,7 @@ public class CourseRepone {
     private String thumbnail;
     private String description;
     private String contents;
-    private float star;
+    private Double star;
     private Double hour;
     private double price;
     private int sold;
